incremental parser is happy with the single-sample rule.
I think when the tests use the parser directly, feeding it a file full
of raw protocol data, the parser needs the whole file described as a
sequence of samples. But the incremental (protocol) parser is okay with
knowing that just the next thing is a sample.

And in fact, trying to treat the streaming protocol data as one long
sequence of samples was running into problems, maybe hitting a buffer
limit.